all movie alone related problems and assignments are completed

+++ src/privatemovie/be/Movie.java
@@ -54,7 +54,7 @@ public class Movie {
     public void setOwnrating(int ownrating) {
         this.ownrating = ownrating;
     }
-    public double getOwnrating(){
+    public int getOwnrating(){
         return ownrating;
     }
